+++ Source/Core/Windows/ThingEditFormUDMF.cs
@@ -198,7 +198,7 @@ namespace CodeImp.DoomBuilder.Windows
 			double floorheight = (ft.Sector != null ? Sector.GetFloorPlane(ft.Sector).GetZ(ft.Position) : 0);
 			posX.Text = (ft.Position.x).ToString();
 			posY.Text = (ft.Position.y).ToString();
-			posZ.Text = (useabsoluteheight ? ((float)Math.Round(ft.Position.z + floorheight, General.Map.FormatInterface.VertexDecimals)).ToString() : (ft.Position.z).ToString());
+			posZ.Text = (useabsoluteheight ? (Math.Round(ft.Position.z + floorheight, General.Map.FormatInterface.VertexDecimals)).ToString() : (ft.Position.z).ToString());
 			posX.ButtonStep = General.Map.Grid.GridSize;
 			posY.ButtonStep = General.Map.Grid.GridSize;
 			posZ.ButtonStep = General.Map.Grid.GridSize;
@@ -271,7 +271,7 @@ namespace CodeImp.DoomBuilder.Windows
 				if((t.Position.y).ToString() != posY.Text) posY.Text = "";
 				if(useabsoluteheight && t.Sector != null) 
 				{
-					if(((float)Math.Round(Sector.GetFloorPlane(t.Sector).GetZ(t.Position) + t.Position.z, General.Map.FormatInterface.VertexDecimals)).ToString() != posZ.Text)
+					if((Math.Round(Sector.GetFloorPlane(t.Sector).GetZ(t.Position) + t.Position.z, General.Map.FormatInterface.VertexDecimals)).ToString() != posZ.Text)
 						posZ.Text = "";
 				} 
 				else if((t.Position.z).ToString() != posZ.Text) 
@@ -610,13 +610,13 @@ namespace CodeImp.DoomBuilder.Windows
 			Thing ft = General.GetByIndex(things, 0);
 			double z = ft.Position.z;
 			if(useabsoluteheight && ft.Sector != null) z += Sector.GetFloorPlane(ft.Sector).GetZ(ft.Position);
-			posZ.Text = ((float)Math.Round(z, General.Map.FormatInterface.VertexDecimals)).ToString();
+			posZ.Text = Math.Round(z, General.Map.FormatInterface.VertexDecimals).ToString();
 
 			foreach(Thing t in things) 
 			{
 				z = t.Position.z;
 				if(useabsoluteheight && t.Sector != null) z += Sector.GetFloorPlane(t.Sector).GetZ(t.Position);
-				string ztext = ((float)Math.Round(z, General.Map.FormatInterface.VertexDecimals)).ToString();
+				string ztext = Math.Round(z, General.Map.FormatInterface.VertexDecimals).ToString();
 				if(posZ.Text != ztext) 
 				{
 					posZ.Text = "";
@@ -720,7 +720,7 @@ namespace CodeImp.DoomBuilder.Windows
 				{
 					double z = posZ.GetResultFloat(thingprops[i++].Z);
 					if(useabsoluteheight && !posZ.CheckIsRelative() && t.Sector != null)
-						z -= (float)Math.Round(Sector.GetFloorPlane(t.Sector).GetZ(t.Position.x, t.Position.y), General.Map.FormatInterface.VertexDecimals);
+						z -= Math.Round(Sector.GetFloorPlane(t.Sector).GetZ(t.Position.x, t.Position.y), General.Map.FormatInterface.VertexDecimals);
 					t.Move(new Vector3D(t.Position.x, t.Position.y, z));
 				}
 			}
